Brutal refactoring. Started contemplating adding eye tracking code as well.
Sorry this commit is both vague and messy - seeing as I'm totally restructuring many things, some files are actually renames not done in git (so lost history...) Justification: the history is more a mess than anything. sTill I know that's a bit lazy but I'm the only one on this at the moment, so nanana.